// vite-project/src/types/UserInfo.ts
export type UserInfo = {
  firstName: any;
  surName: any;
  otherName: any;
  userName: any;
  slug: any;
  gender: any;
  voice: any;
  joined: any;
  graduated: any;
  level: any;
  position: any;
  image: any;
  post: any;
  birthMonthDay: any;
  email: any;
  phone: any;
  password: string;
  token: string;
  isAdmin: boolean;
};