// vite-project/src/types/User.ts
export type User = {
  _id: string;
  firstName: string;
  surName: string;
  otherName: string;
  userName: string;
  gender: string;
  image: any;
  file: any;
  voice: string;
  joined: any;
  graduated: any;
  level: string;
  position: string;
  post: string;
  birthMonthDay : any;
  email: string;
  phone: string;
  password: string;
  token: string;
  isAdmin: boolean;
};